add a patch to build Qt/E 2.3.10 on amd64 with gcc4
(From Opie gcc4_amd64.patch)

+++ src/kernel/qgfxvfb_qws.cpp
@@ -277,7 +277,7 @@ bool QVFbScreen::connect( const QString & )
     else
 	return FALSE;
 
-    if ( (int)shmrgn == -1 || shmrgn == 0 )
+    if ( (long)shmrgn == -1 || shmrgn == 0 )
 	return FALSE;
 
     hdr = (QVFbHeader *) shmrgn;
@@ -307,7 +307,7 @@ bool QVFbScreen::connect( const QString & )
 
 void QVFbScreen::disconnect()
 {
-    if ( (int)shmrgn != -1 && shmrgn ) {
+    if ( (long)shmrgn != -1 && shmrgn ) {
 	shmdt((char*)shmrgn );
 	shmrgn = 0;
     }
